Move the test for cpu implementations of abstractconv over to the gpuarray backend.
